Moved dev map files around somewhat
