+++ src/dev/color-schemes.ts
@@ -1,4 +1,4 @@
-import { VectorMapInspectorPage, MacrostratVectorTileset } from "./map";
+import { VectorMapInspectorPage, MacrostratVectorTileset } from "./map-layers";
 import h from "@macrostrat/hyper";
 import { useState } from "react";
 import { ButtonGroup, Button } from "@blueprintjs/core";
